:zap:feat:ver6,7

+++ src/main/java/org/example/plan/service/MemberService.java
@@ -2,10 +2,13 @@ package org.example.plan.service;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.example.plan.config.PasswordEncoder;
 import org.example.plan.dto.MemberResponseDto;
 import org.example.plan.dto.SignUpResponseDto;
 import org.example.plan.entity.Member;
+import org.example.plan.entity.Plan;
 import org.example.plan.repository.MemberRepository;
+import org.example.plan.repository.PlanRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.util.PatternMatchUtils;
@@ -19,15 +22,12 @@ import java.util.Optional;
 public class MemberService {
 
     private final MemberRepository memberRepository;
-    private static final String[] emailList = {"*@*","*@*"};
+    private final PasswordEncoder passwordEncoder;
+    private final PlanRepository planRepository;
 
     public SignUpResponseDto signUp(String username, String email, String password) {
 
-        if(!isWhiteList(email)){
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "이메일 형식에 맞게 입력하세요.");
-        }
-
-        Member member = new Member(username, email, password);
+        Member member = new Member(username, email, passwordEncoder.encode(password));
         Member saveMember = memberRepository.save(member);
 
         return new SignUpResponseDto(saveMember.getId(),saveMember.getUsername(),saveMember.getEmail(),saveMember.getPassword());
@@ -62,11 +62,12 @@ public class MemberService {
 
     public void delete(Long id){
         Member findMember = memberRepository.findByIdOrElseThrow(id);
+        Plan findPlan = planRepository.findByIdOrElseThrow(id);
+
+        if (!findPlan.getTitle().isEmpty()) {
+            throw new IllegalStateException("사용자의 게시글이 존재합니다. 삭제할 수 없습니다.");
+        }
 
         memberRepository.delete(findMember);
     }
-
-        private boolean isWhiteList(String email) {
-            return PatternMatchUtils.simpleMatch(emailList, email);
-        }
 }
